code cleaning & simplification
git-svn-id: http://svn.miranda-ng.org/main/trunk@6138 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/JabberG/src/jabber_thread.cpp
@@ -2041,7 +2041,7 @@ int ThreadData::send(char* buffer, int bufsize)
 		return 0;
 
 	if (bufsize == -1)
-		bufsize = strlen(buffer);
+		bufsize = (int)strlen(buffer);
 
 	WaitForSingleObject(iomutex, 6000);
